Now handles file deletion in the PinPhoto class, rather than in the destroyer class. This is done by request of the rubric, not because I think it's a good practice.

+++ VirtualTourist/VirtualTourist/AlbumDestroyer.swift
@@ -13,6 +13,10 @@ import PromiseKit
 public class AlbumDestroyer {
     let context: NSManagedObjectContext
     let organizer: PhotoOrganizer
+    /** The rubric wants us to handle this file deletion in the model class
+    so this flag lets us halt the file deletion we already wrote here.
+    */
+    var shouldDestroyPhotoFilesWhenDestroyingPin = false
     init(context c: NSManagedObjectContext, organizer o: PhotoOrganizer) {
         self.context = c
         self.organizer = o
@@ -25,18 +29,20 @@ public class AlbumDestroyer {
         // Delete all of the photos from the file system
         // This could be done asynchronously as not to lock 
         // up the main thread, even if it only takes a moment
-        var paths = [String]()
-        for item in pin.photos {
-            if let photo = item as? PinPhoto {
-                paths.append(organizer.path(photo.fileName))
+        if !shouldDestroyPhotoFilesWhenDestroyingPin {
+            var paths = [String]()
+            for item in pin.photos {
+                if let photo = item as? PinPhoto {
+                    paths.append(organizer.path(photo.fileName))
+                }
             }
-        }
-        // If we do this asynchronously the record will probably be 
-        // deleted/deleting while we're still going through the files
-        // Hence, the copied list of paths
-        dispatch_async(dispatch_get_main_queue()) { () -> Void in
-            for path in paths {
-                self.organizer.delete(path)
+            // If we do this asynchronously the record will probably be
+            // deleted/deleting while we're still going through the files
+            // Hence, the copied list of paths
+            dispatch_async(dispatch_get_main_queue()) { () -> Void in
+                for path in paths {
+                    self.organizer.delete(path)
+                }
             }
         }
         // Deleting the Pin should cascade deletions of all the PinPhotos
@@ -50,7 +56,9 @@ public class AlbumDestroyer {
     
     public func destroy(photos: [PinPhoto]) throws {
         for i in photos {
-            destroyFile(i)
+            if shouldDestroyPhotoFilesWhenDestroyingPin {
+                destroyFile(i)
+            }
             context.deleteObject(i)
         }
         try context.save()
